[S0007_legendre_stress_test] Commit stress-testing program.
This was designed to easily consume CPU cycles and thrash the CPU caches so
that the Intel Haswell machines could be evaluated. The group is considering
buying a dedicated machine, and there are questions about whether the extra
cost of the Intel machines are worth it when compared to computation times on
AMD.

The issue is that the interactive nodes are often under significant load, so
doing a performance comparison against this test Intel system which isn't
being used except by me is unfair. To level the playing field, I've chosen
to synthetically load the system. This program does that.